Refactor code for a cleaner codebase
