+++ .pre-commit-config.yaml
@@ -30,7 +30,7 @@ repos:
         language_version: python3.10
         args: ['--target-version', 'py310']
   - repo: https://github.com/asottile/pyupgrade
-    rev: v2.19.0
+    rev: v3.4.0
     hooks:
       - id: pyupgrade
         args: [--py37-plus]
