+++ README.md
@@ -28,4 +28,4 @@ If the script fails, please check out the [setup doc](./docs/setup.md) for instr
 ## Running The Program
 
 Once you are done with dev setup. Enter the virtual environment with `source .venv/bin/activate` if you are on a Mac/Unix/Linux system, or with `.\venv\Scripts\activate` if you are on a Windows system.
-Then simply run `python driver.py` to start the program.
+Then simply run `python main.py` to start the program.
