--- driver.py
@@ -1,8 +1,0 @@
-#!/usr/bin/python
-# Author: Hoang Truong
-from src.home_window import HomeWindow
-
-
-if __name__ == "__main__":
-    while True:
-        HomeWindow()
