+++ main.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python
+# Author: Hoang Truong
+from src.home_window import HomeWindow
+
+
+if __name__ == "__main__":
+    while True:
+        HomeWindow()
