+++ requirements-dev.txt
@@ -1,1 +1,10 @@
+cfgv==3.3.1
+distlib==0.3.6
+filelock==3.12.0
+identify==2.5.23
+nodeenv==1.7.0
+platformdirs==2.6.2
 pre-commit==2.20.0
+PyYAML==6.0
+toml==0.10.2
+virtualenv==20.16.5
