+++ src/data_structures/__init__.py
@@ -1,0 +1,5 @@
+from .priority_queue import PriorityQueue
+from .queue import Queue
+from .stack import Stack
+
+__all__ = ["PriorityQueue", "Queue", "Stack"]
