+++ src/visualizers/__init__.py
@@ -1,0 +1,4 @@
+from .pathfinding.pathfinding_visualizer import PathfindingVisualizer
+from .sorting.sorting_visualizer import SortingVisualizer
+
+__all__ = ["PathfindingVisualizer", "SortingVisualizer"]
