+++ src/visualizers/sorting/__init__.py
@@ -1,0 +1,3 @@
+from .sorting_visualizer import SortingVisualizer
+
+__all__ = ["SortingVisualizer"]
